scripts/requirements.txt
jinja2
markdown-it-py[linkify,plugins]
python-frontmatter